ci: Add support for building el8stream images
This patch adds the pipeline definition to build CentOS Stream
images.

The change is mostly implemented by:
- defining the pipeline in 'stdci.yaml'
- defining the necessary variables for building el8stream
  in 'automation/ost-images-el8stream.environment.yaml'

The logic of the makefile/scripts is not changed.

Both kickstart file and provisioning scripts are reused from el8.

With this change, the target which builds el8 images, that was called
'ost-images' earlier, is now renamed to 'ost-images-el8'.

Change-Id: Iaa1a146ef9a4f8f7720b34b4f7a58703c7cde020
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>