ci: Added '*-no-base' pipelines
This patch adds '-no-base' pipelines for el8 and el8stream.

The pipelines build the 'upgrade' layer and the '{engine,host}-installed'
layers, but use the prebuilt 'base' layers by simply installing them
from 'templates.ovirt.org/yum' repo.

These pipelines are handy, since we don't need to rebuild the base layer
most of the times, so the running time is vastly shortened. The jobs that
run nightly will be eventually switched to use them.

There is no change in the logic of the shell scripts or the Makefile
for those pipelines - the gist of the implementation is adding
the 'ost-images-{el8,el8stream}-base' dependencies to the appropriate
'automation/*.packages' file - Makefile logic [1] handles the rest.

[1] https://gerrit.ovirt.org/gitweb?p=ost-images.git;a=blob;f=Makefile.am;h=26a01ae523b4b5d494414edadd0fa83c9d9259c5;hb=refs/heads/master#l16

Change-Id: Ia702b8c3233619818e923a235971359f4f9e1f90
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>