make: Separate makefile for built/preinstalled base
This patch extracts the definitions of the base-layer-related
targets to 'build-base.mk' and 'preinstalled-base.mk' sub-makefiles.

A nice side effect of this change is that now we can use the same
'%-base.qcow2' target name in both built, and preinstalled base cases
instead of ugly '%-preinstalled-base', which was ultimately renamed
to '%-base.qcow2'.

Change-Id: I94e38cedf212bc4bceaf5784fb6a758332bde303
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>