make: Support using preinstalled upgrade layer
This patch adds the possibility to use preinstalled 'upgrade'
layer to build only '{engine,host}-installed' layers.

The use case for this functionality is to quickly build functional
VM images from a set of custom RPMs for the engine or the host.

The mechanics are almost identical to the ones already used
for building without the base layer.

Change-Id: I2b63f19f1d5e38783d22e51add836533403c9de2
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>