make: Make building '{engine,host}-installed' RPMs optional
This patch adds the possibility to skip building either 'engine-installed',
or 'host-installed' layer, but not both at once.

The use case for this feature is to build a single VM image
for a set of RPMs to be tested, i.e. if you want to test
a 'vdsm' build, you don't really need to build
the 'engine-installed' layer.

The way this feature works, is we have 2 text files:

- tested-engine-packages.txt
- tested-host-packages.txt

and both of these files contain plain names of {engine,host}-related packages
(i.e. 'vdsm', 'ovirt-engine'). We also have the 'EXTRA_REPOS' variable,
which can contain a bunch of URLs. These two are glued together using
'find-packages-in-repo.sh' script. The script simply uses 'curl' to download
the index page of each repository passed in 'EXTRA_REPOS' and grep for the
listed packages. More or less:

 curl http://jenkins.ovirt.org/first/repo/url | egrep 'ovirt-engine|ovirt-smth|smth-else'
                                                       ^
 This contains all packages from the txt file _________|

if a match is found, we assume that the build for certain layer is desired.

Change-Id: Idcd96f956ecc8e70f6b05707921c218d47ab1fd4
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>