make: Introduce '{engine,host}-deps-installed' layers
This patch adds two more layers to the picture -
'engine-deps-installed' and 'host-deps-installed'.

The images contain as much dependencies for the 'ovirt-engine'
and 'ovirt-host' packages as possible, but nothing that comes from
'ovirt' repositories.

The rationale behind these layer is most of the time we want to test
_fresh installations_ of our packages. If a user works on a ready-made
'host-installed' layer, and puts his own 'vdsm' build on top of that,
he/she will be doing an _upgrade_ of the vdsm package, not a fresh
installation. This is meaningful for some packages since RPM scripts
vary between those two scenarios.

The way the layers are built is we:

- add any 'EXTRA_REPOS' passed to the VM
- add the 'ovirt-release-rpm'
- do any other preparations (like enabling some dnf modules for the engine)
- download the 'ovirt-{engine,host}' RPM with all the dependencies
  to a temporary folder
- use 'tested-{host,engine}-packages.txt' file to filter out anything
  ovirt-related from the downloaded RPMs
- install remaining RPMs with '--skip-broken' and '--disable-repo *'
  options not to pull anything ovirt-related accidentally again

Change-Id: I748413fc1c1325c5a681093096bd8a55b0e34714
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>