host: Move 'python3-coverage' installation to happen later
The 'python3-coverage' that's available throught standard repo
is weird - it contains mostly the '/usr/bin/coverage-3' script,
while the usual 'site-packages' stuff is provided by the
'platform-python-coverage' package, which the former depends on.
Unfortunately this causes file-level conflicts with 'python3-coverage'
that's provided by Sandro's repos, which gives both the script
and the 'site-packages' stuff.

To avoid the conflict we're moving the 'python3-coverage' installation
after adding all the release RPM repos, and hence, using the version
from Sandro's repo.

Change-Id: Iccbca07f05b690c59241784c4365461122c3a6eb
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>