make: Drop '*deps-installed*' layers
The layers that had engine/host dependencies installed
were meant to provide a quick way for having a completely clean
installation of the engine/vdsm package.

We don't use them currently and they complicate things a lot.
If at any point in the future we'll need the kind of functionality
they provided, we can reintroduce them.

Change-Id: I82154eeb6849e4b79a8d931d9026f5a99a5926e7
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>