he: Add new layer with ovirt-engine-appliance installed
This patch introduces another layer to ost-images -
'${distro}-he-installed'. It's based on '${distro}-host-installed' layer
and adds 'ovirt-engine-appliance' package on top of that.

Using this image we can omit the whole process of downloading
and installing the engine appliance [1] each time we run any of the
hosted engine suites, which saves quite a lot of time.

[1] https://github.com/oVirt/ovirt-system-tests/blob/master/common/deploy-scripts/install_he_4.sh

Change-Id: I077e405c5973ba678c7d791b87b9bff7b652303f
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>