Reduce all images' sizes
This patch introduces 2 optimizations to reduce sizes of the built
qcows and RPMs.

When you i.e. create a 1MB file inside the qcow during customization,
the qcow will grow by 1MB. When you remove the file however, the qcow
doesn't shrink automatically. Freeing up the space used by removed
files can be done with 'virt-sparsify' command, so we use it on the
'${distro}-base' layer.

Unfortunately the 'virt-sparsify' command flattens the images,
so we can't really use it on layers other than the 'base' one.
To achieve size reduction on the lower layers we use a different
strategy - we use '/dev/shm' as cache directory for dnf.
This makes all downloaded packages to magically disappear after
installation without using any space on the built qcow image.

The approximate numbers are:

layer                 qcow size reduction    RPM size reduction
---------------------------------------------------------------
base                        16%                   34%
engine-installed            27%                   50%
host-installed              20%                   52%
he-installed                50%                   50%

The downside is we need much more RAM to build those qcows.
The 'ovirt-engine-appliance' package used in '${distro}-he-installed'
layer is the biggest one - ~2GB in size. This is why we need to make
'virt-customize' command to reserve 5120MBs of RAM for the customization
process.

Change-Id: I1d8ce51b71b4dcbe6a18ff9964c859105dc87897
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>