rpm: Use multiple threads when compressing
It seems that when building ost-images RPMs the bottleneck
is in the xz compression stage - on a decent CPU (i7@3.40GHz)
and very fast storage (DDR3 ramdisk) we're talking about speeds ~3MB/s.
The xz compression that we use gives great results when it comes
to ratio, but is indeed known for being slow.

Fortunately, there is a way to speed up this process. A quick analysis
shows that, by default, only 1 thread is running the compression
algorithm, but RPM spec syntax allows passing an option to define
how many threads you want to use. In my tests, the improvement has
shown to be close to linear.

This patch adds a 'XZ_NUM_THREADS' variable to the makefiles, which
can be overriden as other variables, i.e.:

 make XZ_NUM_THREADS=5 rpm

or:

 export XZ_NUM_THREADS=5
 make -e rpm

The variable is passed to the spec to parametrize '_source_payload' and
'_binary_payload' variables, which define the compression parameters.
The default value has been set to 2.

When overriding with your own values beware that during compression
the threads eat up almost 100% of the CPU all the time.

Change-Id: Ia98f386edcfa1e637a2143b2d147581246d85565
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>