base: Use static ssh keys
Until now when we were generating any base image and ssh key was
dynamically created and added to the qcow and the RPMs.
When testing oVirt node we're using 2 different base images -
the engine is el8/el8stream/rhel8-based while node is a base image
on itself. This is problematic on OST side were we can pass only 1 key
to lago [1][2] and it's not easy to get around that. The obvious
solution for now is to include the same set of ssh keys to all bases
built here.

[1] https://github.com/oVirt/ovirt-system-tests/blob/904a862217cf92c19cd894b322b81e67fac13d85/run_suite.sh#L160
[2] https://github.com/oVirt/ovirt-system-tests/blob/904a862217cf92c19cd894b322b81e67fac13d85/common/helpers/ost-images.sh#L14

Change-Id: I8d5a945080fd74f1b75bdd7e34f03fba5b6e3188
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>