ks: Make VMs use hvc console instead of serial
Without this patch:

lago creates VMs with both serial and hvc, but serial is not connected
anywhere on the host (no "source="), AFAICT, so can't be accessed.
lago/virsh hvc console gets a getty but no kernel messages.

This patch tells the engine to use hvc0 as the console, so that we get
there kernel output.

I am pushing this patch to help debug an issue with host VMs getting
stuck after/during reboot.

This probably isn't enough for also getting grub output. If it seems
like a bios/grub issue, another patch (here or for lago) will be needed.

Change-Id: I76353c9ead89a743806f7f9d5d10272f354fc978
Signed-off-by: Yedidyah Bar David <didi@redhat.com>