ks: Clean dnf cache in post install
When building images from i.e. compose URLs we leave dnf
metadata behind in the image. This causes issues when trying
to run OST with the same compose URL provided as "extra repository" -
dnf usually thinks the existing metadata is fresh enough,
so it's not updated and new packages are not being picked up.

Change-Id: I986880c7043c5410da4fd0ce63be2a6c124e086f
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>