base: Add stable ipv6 drop-in config for NM
During OST setup we add a drop-in configuration file [1][2]
that determines the configuration of ipv6 addresses.
After that is done we restart the NetworkManager, receive
the ipv6 addresses reserved by lago from dnsmasq and assign
them to the NICes.

The problem is when that happens we already possess an ipv6
address for each of the NICes assigned at boot time. The effect
is we end up with having 2 ipv6 addresses per NIC.

This is problematic for other deployment scripts which may want
to use ipv6 addresses as well - some of them may get the "boot time"
one and some the "dhcp" one.

Let's bake-in the NM config file to the base image - that will give
us a guarantee that we will always have only one ipv6 address per NIC
assigned by the dhcp server.

[1] https://github.com/oVirt/ovirt-system-tests/blob/67202bdcd6bf4a16cb69f94cb4d76c2225ef264b/common/deploy-scripts/setup_host.sh#L57
[2] https://github.com/oVirt/ovirt-system-tests/blob/67202bdcd6bf4a16cb69f94cb4d76c2225ef264b/common/deploy-scripts/setup_engine.sh#L76

Change-Id: I24d66bf34db5b7f9061dafbf88e80642b80d90ac
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>