Remove support for 'EXTRA_REPOS' variable
Until now we had logic in the build process that allowed
the user to pass some extra repositories that would be used when
building '{host,engine}-installed' layers of ost-images.
We don't really use it as we simply pass the extra repos in the OST
run itself and run 'dnf upgrade' then.

Change-Id: I14b5829ed00a891f8c59f3c3ef8abf2ac4378f25
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>