introduce provision-base.sh
We now have few things that are added in base image already, let's have
a common provision-base.sh that runs right after the image is installed.
It's intended for things that are common to every ost-image, ideally
these are distro-agnostic so there's no need for a per-distro script
right now.

Change-Id: Ie02320ef78e1a38d260217e60f108407a1a29716
Signed-off-by: Michal Skrivanek <michal.skrivanek@redhat.com>