Adding tr-ocp-installer package
Adding tr-ocp-installer package.

This is a dummy package that will have
the real openshift-install bin just after upgrade.

The openshift-install bin is built with the related patch in
terraform-provider-ovirt repo and its purpose is to check if changes
suggested by the patch are fully compatible with OCP.

In its initial form the package contains a shell script
named openshift-install that prints an error and exists.

Signed-off-by: emesika <emesika@redhat.com>