base: Remove support for builds based on top of existing base
Until now we had the option to use an existing base image
(one installed along 'ost-images-DISTRO-base' RPM package)
as the foundation for all the other images. This made sense
when we were building CentOS 8 images with 'upgrade' layer
in between, but now we're working with rolling release distros
or nightly repos. Let's remove the logic that allowed that.

Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>