make: Use the templating mechanism we already have
We always had the provisioning scripts as '*.in' files,
so we can adjust them in build time. Until now we didn't
really do anything special with them - the '.in' file were
simply copied to regular ones. We do need the templating
for the rhel8 provisioning scripts though.

Let's use the mechanism we always had instead of making
the substitutions on 'build.sh' level.

Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>