Introducing 'storage' image
This patch introduces a new kind of "base" image called "storage".

It's main purpose will be hosting storage domains for existing suites
(instead of having them on the engine VM) and providing Selenium
Grid functionality.

The Selenium Grid is not up upon boot. This is because video containers
start recording immediately when they're started. To use the grid
one needs to bring up the 'selenium-pod' systemd service with:

 systemctl start selenium-pod

The grid is exposed on port 4444.

Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>