el9node: fix FIPS setup during installation
it doesn';t look like FIPS setup is completed properly on el9stream
node. IT could be a recent change as it used to pass before, but now it
reports an inconsistent state, not having "fips=1" on kernel cmdline. It
may be yet another thing to run in RW layer during installation.