jenkins timing issue
due to the way jenkins jobs are chained there's a timing issue with
archiving the results, if we run a new build right after the previous
one the artifacts may not have been copied just yet. It's idiotic, but
let's just wait a little...2 minutes ought to be enough for anyone.