ci: various changes to the github pipeline
- don't run the container/pipeline as build user anymore.
actions/upload-artifact is broken under non-root user:
https://github.com/actions/upload-artifact/issues/542

- fix libvirt under cgroupsv2 by adding --cgroupns host
https://github.com/moby/moby/issues/42275

- start libvirtd and virtlogd
We need to start libvirtd and virtlogd in order to make virt-install
working inside the container.

- checkout
use actions/checkout as the workaround is not needed anymore as we run
as root anyway

- logs
Append os name to the logs so we don't have duplicate artifacts.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>