build: temporarily disable fips
In RHEL 9, which ships with OpenSSL 3.0, a critical compatibility issue
arises during PKCS#12 keystore generation.

From OpenSSL 3.0 onwards, the default algorithms for PKCS#12 generation
have been updated to more modern and secure standards (e.g., PBES2 with
AES-256 and SHA-256). Simultaneously, OpenSSL 3.0's FIPS provider explicitly
excludes certain legacy operations like the PKCS12KDF, preventing it from
generating PKCS#12 files with encryption schemes that are compatible with
Red Hat's FIPS-enabled OpenJDK 11.

This fundamental incompatibility breaks the encryption setup of
the ovirt-engine, leading to failures in several system tests.

In Java 21 the support for PBES2 does exist. So, as we plan to move to
Java 21 in the near future, we will disable FIPS in the ost-images for
now, and re-evaluate the usage of FIPS after the migraton to java 21.

Signed-off-by: Peter Boden <peter.boden@team.blue>